requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
#psycopg2==2.9.2
gunicorn==20.1.0 
python-dotenv==0.19.1
pytz